<<Finished upfront work for realtime notifications>>
Summary: Fixed the server to manage connections more accurately

Test Plan:
Opened up a few windows, put two on the same page,
one on a different page.  commented.  The other person on the page
was told to refresh.  Nobody else was.
I think it's best that we all just use it for a while
and when it messes up check why.  I've been running the server
while changing my php code and recommenting and it seems to be fine.

Reviewers: david

Reviewed By: david

CC: keebuhm, epriestley

Differential Revision: http://theoryphabricator.com/D3