Fix for logging into /notifications
Summary: This is a fix for the bug where logging in leads users to /notifications page. It checks $request->getUser()->isLoggedIn() to check if the user of the request is logged in. If not, the header does not have the notification indicator or the panel.

Test Plan: Logging in now takes you to the main page, not /notifications

Reviewers: allenjohna, david

Reviewed By: david

Maniphest Tasks: T146

Differential Revision: http://theoryphabricator.com/D22