Adds base notification application
Summary: First diff in a series of diffs to add notifications to Phabricator. This is the notification application ONLY. This commit does not include the changes to other applications that makes them add notifications. As such, no notifications will be generated beyond the initial database import.

Test Plan: This is part of the notifications architecture which has been running on http://theoryphabricator.com for the past several months.

Reviewers: epriestley, btrahan

CC: allenjohnashton, keebuhm, aran, Koolvin

Maniphest Tasks: T974

Differential Revision: https://secure.phabricator.com/D2571